# WIP: Strict schema validation for POST and PUT