--- a/staging/src/k8s.io/apiserver/pkg/endpoints/apiserver_test.go
+++ b/staging/src/k8s.io/apiserver/pkg/endpoints/apiserver_test.go
@@ -4351,7 +4351,7 @@ func TestUpdateChecksAPIVersion(t *testing.T) {
 	}
 }
 
-// runRequest is used by TestDryRun and TestStrictValidation since it runs the test
+// runRequest is used by TestDryRun and TestFieldValidation since it runs the test
 // twice in a row with a slightly different URL (one has ?dryRun, one doesn't).
 func runRequest(t *testing.T, path, verb string, data []byte, contentType string) *http.Response {
 	request, err := http.NewRequest(verb, path, bytes.NewBuffer(data))
@@ -4404,7 +4404,7 @@ func (storage *SimpleRESTStorageWithDeleteCollection) DeleteCollection(ctx conte
 	return nil, nil
 }
 
-func TestStrictValidation(t *testing.T) {
+func TestFieldValidation(t *testing.T) {
 	strictDecoderErr := "strict decoder error for"
 	// TODO: add test cases for yaml validation and protobuf early exit.
 	tests := []struct {
@@ -4448,7 +4448,7 @@ func TestStrictValidation(t *testing.T) {
 		if response.StatusCode == http.StatusBadRequest {
 			t.Fatalf("unexpected BadRequest: %#v", response)
 		}
-		response = runRequest(t, baseURL+test.path+"?validate=strict", test.verb, test.data, test.contentType)
+		response = runRequest(t, baseURL+test.path+"?fieldValidation=Strict", test.verb, test.data, test.contentType)
 		buf := new(bytes.Buffer)
 		buf.ReadFrom(response.Body)
 		// TODO: better way of doing than string comparison since we are getting a response instead of a regular go error?
@@ -4459,45 +4459,14 @@ func TestStrictValidation(t *testing.T) {
 
 }
 
-// TODO: this is a pretty crude benchmark since it only operates on the simples resource
-// I imagine we want to benchmark some meatier resources too, because maybe the discrepancy in
-// performance is greater for bigger objects?
-func BenchmarkNonStrictValidationSimples(b *testing.B) {
-	server := httptest.NewServer(handle(map[string]rest.Storage{
-		"simples": &SimpleRESTStorageWithDeleteCollection{
-			SimpleRESTStorage{
-				item: genericapitesting.Simple{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "id",
-						Namespace: "",
-						UID:       "uid",
-					},
-					Other: "bar",
-				},
-			},
-		},
-		"simples/subsimple": &SimpleXGSubresourceRESTStorage{
-			item: genericapitesting.SimpleXGSubresource{
-				SubresourceInfo: "foo",
-			},
-			itemGVK: testGroup2Version.WithKind("SimpleXGSubresource"),
-		},
-	}))
-	defer server.Close()
-
-	for n := 0; n < b.N; n++ {
-		postPath := "/namespaces/default/simples"
-		postData := []byte(`{"kind":"Simple","apiVersion":"test.group/version","metadata":{"creationTimestamp":null},"other":"bar"}`)
-		basePostURL := server.URL + "/" + prefix + "/" + testGroupVersion.Group + "/" + testGroupVersion.Version
-		_ = runRequestOrDie(basePostURL+postPath, "POST", postData, "")
-
-		putPath := "/namespaces/default/simples/id"
-		putData := []byte(`{"kind": "Simple", "apiVersion": "test.group/version", "metadata": {"name": "id", "creationTimestamp": null}, "other": "bar", "unknown": "baz"}`)
-		basePutURL := server.URL + "/" + prefix + "/" + testGroupVersion.Group + "/" + testGroupVersion.Version
-		_ = runRequestOrDie(basePutURL+putPath, "POST", putData, "")
+func BenchmarkFieldValidation(b *testing.B) {
+	benchmarks := []struct {
+		name       string
+		queryParam string
+	}{
+		{"nonstrict simples", ""},
+		{"strict simples", "?fieldValidation=Strict"},
 	}
-}
-func BenchmarkStrictValidationSimples(b *testing.B) {
 	server := httptest.NewServer(handle(map[string]rest.Storage{
 		"simples": &SimpleRESTStorageWithDeleteCollection{
 			SimpleRESTStorage{
@@ -4520,16 +4489,21 @@ func BenchmarkStrictValidationSimples(b *testing.B) {
 	}))
 	defer server.Close()
 
-	for n := 0; n < b.N; n++ {
-		postPath := "/namespaces/default/simples"
-		postData := []byte(`{"kind":"Simple","apiVersion":"test.group/version","metadata":{"creationTimestamp":null},"other":"bar"}`)
-		basePostURL := server.URL + "/" + prefix + "/" + testGroupVersion.Group + "/" + testGroupVersion.Version
-		_ = runRequestOrDie(basePostURL+postPath+"?validate=strict", "POST", postData, "")
+	for _, bm := range benchmarks {
+		b.Run(bm.name, func(b *testing.B) {
+			for n := 0; n < b.N; n++ {
+				postPath := "/namespaces/default/simples"
+				postData := []byte(`{"kind":"Simple","apiVersion":"test.group/version","metadata":{"creationTimestamp":null},"other":"bar"}`)
+				basePostURL := server.URL + "/" + prefix + "/" + testGroupVersion.Group + "/" + testGroupVersion.Version
+				_ = runRequestOrDie(basePostURL+postPath+bm.queryParam, "POST", postData, "")
 
-		putPath := "/namespaces/default/simples/id"
-		putData := []byte(`{"kind": "Simple", "apiVersion": "test.group/version", "metadata": {"name": "id", "creationTimestamp": null}, "other": "bar", "unknown": "baz"}`)
-		basePutURL := server.URL + "/" + prefix + "/" + testGroupVersion.Group + "/" + testGroupVersion.Version
-		_ = runRequestOrDie(basePutURL+putPath+"?validate=strict", "POST", putData, "")
+				putPath := "/namespaces/default/simples/id"
+				putData := []byte(`{"kind": "Simple", "apiVersion": "test.group/version", "metadata": {"name": "id", "creationTimestamp": null}, "other": "bar", "unknown": "baz"}`)
+				basePutURL := server.URL + "/" + prefix + "/" + testGroupVersion.Group + "/" + testGroupVersion.Version
+				_ = runRequestOrDie(basePutURL+putPath+bm.queryParam, "PUT", putData, "")
+			}
+		})
+
 	}
 }