… test/integration/apiserver/testdata/pod-large.json
… test/integration/apiserver/testdata/pod-medium.json