@kevindelgado 2021-08-18
Added some benchmarking per our discussion from the sig/apimachinery meeting @lavalamp.

Let me know if you think there’s any issues with the methodology (I’ve noted that we’re only benchmarking one simple resource), otherwise it looks like doing the strict validation is about 20% slower than not doing it. 

The point of the benchmarking was to determine whether we want to do strict validation only when requested, or if we should do it on every request and warn instead of error when the user doesn’t want strict validation. What do you think now given these results? 

```
BenchmarkNonStrictValidationSimples-16              1236            858508 ns/op
BenchmarkNonStrictValidationSimples-16              1261            847371 ns/op
BenchmarkNonStrictValidationSimples-16              1263            844816 ns/op
BenchmarkNonStrictValidationSimples-16              1345            887105 ns/op
BenchmarkNonStrictValidationSimples-16              1393            979278 ns/op

BenchmarkStrictValidationSimples-16                 1238           1054536 ns/op
BenchmarkStrictValidationSimples-16                 1358            954324 ns/op
BenchmarkStrictValidationSimples-16                 1251           1171214 ns/op
BenchmarkStrictValidationSimples-16                 1122           1152582 ns/op
BenchmarkStrictValidationSimples-16                 1074           1193945 ns/op
```