@apelisse 2021-08-18
You can possibly use [benchstat](https://pkg.go.dev/golang.org/x/perf/cmd/benchstat) to display these results, and you probably should include [memory usage](https://pkg.go.dev/testing#B.ReportAllocs) too