@apelisse 2021-08-18
If we wanted to have these as warnings, we would have to run twice then. We would have to decide in what order (probably strict first and then non-strict, since strict shouldn't fail in most cases). The nominal case would then be to do it mostly once with strict.

Something we could do is to try to run this against a huge cluster and see if it has any impact on latency.