@smarterclayton 2021-09-03
I don't understand why strict serializer is so expensive.  While 20-30% is acceptable for opt in, the actual checks should be amortizable (we already have to iterate every field before assignment).  Naively I would expect that strict serialization should be capable of being free (<3-5%)