@lavalamp 2021-09-07
> I don't understand why strict serializer is so expensive

Judging by sub-ms times, I think "simples" are very tiny and any constant overhead is going to be pronounced. Can we try with bigger objects?