@kevindelgado 2021-09-14
> > I don't understand why strict serializer is so expensive
> 
> Judging by sub-ms times, I think "simples" are very tiny and any constant overhead is going to be pronounced. Can we try with bigger objects?

@lavalamp This has sent me down a bit of a rabbit hole, so I wanted to check in before I go too deep. 

I’m trying to benchmark objects bigger than `simples`. The place that I was benchmarking the simples `apiserver/pkg/endpoints/apiserver_test.go`, doesn’t seem like an appropriate place to benchmark bigger objects, because you have to manually add the rest.Storage map to the test server (and just from a consistency perspective, nothing else in this package uses any objects beyond `simples`).

My approach has instead been to [add a benchmark](https://github.com/kubernetes/kubernetes/pull/104433/files#diff-c5e700b1b521973df191a7218a45a6aecaa4081af17a43f9bfbca2b9a287cf0eR2333) to the integration test in `test/integration/apiserver/apiserver_test.go`since this gives us access to the full suite of resources.

This has given promising results with strict validation being ~11% slower and ~25% more memory. 

```
❯ benchstat benchvalidation-10x-nonstrict.log benchvalidation-10x-strict.log
name                                  old time/op    new time/op    delta
FieldValidation/strict-deployment-16    5.18ms ± 5%    5.73ms ± 9%  +10.76%  (p=0.001 n=9+9)
 
name                                  old alloc/op   new alloc/op   delta
FieldValidation/strict-deployment-16     264kB ± 0%     323kB ± 1%  +22.47%  (p=0.000 n=10+10)
 
name                                  old allocs/op  new allocs/op  delta
FieldValidation/strict-deployment-16     3.23k ± 0%     4.07k ± 0%  +26.04%  (p=0.000 n=10+10)
```

Is this a better place to do the benchmarking? Are there other/better places to add benchmarking? I also considered adding benchmarking to the low level `serializer/json` package but that has its own issues (doesn’t seem like it would benchmark anything that jsoniter wouldn’t already do, doesn’t actually capture the effect on the whole API request, etc).

If this is a good place to put it, I can go ahead and add even bigger objects to the testing table so that we can see the effect of object size on slowness of strict validation. (Aside: do you know where I can find objects of various sizes in raw json/yaml that I can test here, otherwise I can hand craft them myself, but wanted to check first?).